+++ pkg/log/messages/debug_messages.go
@@ -8,3 +8,4 @@ const CSPFK006D string = "CSPFK006D Kubernetes Secret '%s' has no value defined 
 const CSPFK007D string = "CSPFK007D Kubernetes Secret '%s' has an invalid value for '%s' data entry. Reason: %s"
 const CSPFK008D string = "CSPFK008D Kubernetes Secret '%s' has no '%s' data entry defined"
 const CSPFK009D string = "CSPFK009D Processing '%s' data entry value of Kubernetes Secret '%s'"
+const CSPFK010D string = "CSPFK010D Listed %d secrets from Conjur"