+++ pkg/log/messages/error_messages.go
@@ -43,6 +43,7 @@ const CSPFK025E string = "CSPFK025E PathMap cannot be empty"
 const CSPFK027E string = "CSPFK027E Failed to update Kubernetes Secrets map with DAP/Conjur secrets"
 const CSPFK028E string = "CSPFK028E Unable to update Kubernetes Secret '%s'"
 const CSPFK063E string = "CSPFK063E Unable to delete Kubernetes Secret"
+const CSPFK067E string = "CSPFK067E Normalized secret names produced a duplicate secret key '%s'. Leaving first value in place."
 
 // DAP/Conjur
 const CSPFK031E string = "CSPFK031E Failed to load DAP/Conjur config. Reason: %s"
@@ -88,3 +89,7 @@ const CSPFK061E string = "CSPFK061E Could not write content to temporary file fo
 // Secrets Decoding
 const CSPFK064E string = "CSPFK064E Failed to decode secret '%s' with '%s' content-type. Reason: %s"
 const CSPFK065E string = "CSPFK065E Secret group %s: the content-type of %s is invalid"
+
+// Fetch All
+const CSPFK066E string = "CSPFK066E Stopped fetching additional secrets after %d secrets were fetched"
+const CSPFK068E string = "CSPFK068E Retrieved secrets did not include secret '%s' requested by secret group '%s'"